requirements.txt
Django==2.2.1
django-model-utils==3.1.2
djangorestframework==3.9.4
Pillow==6.0.0
psycopg2-binary==2.8.2
django-storages==1.7.1
django-cors-headers==3.0.2
django-object-actions==1.1.0
